modify the code snippet to check all analysis result has been calculated
the bug is because I pick up the result between long point and short point to calculate when I calculate the analysed result. However, a period at the tail can't fit this condition, so the analysed result has nothing changed (still 0's). Therefore, I modify this bug w/ the method which get the max distance value at that period out of the loop